+++ src/utilities/extensions.rs
@@ -4,16 +4,6 @@ use chrono::{DateTime, NaiveDate, NaiveDateTime, Utc};
 use tracing::error;
 
 pub trait CustomStringExtensions {
-    /// ## Check if the value is a valid json key for hubuum
-    ///
-    /// We support only ASCII alphanumeric characters, underscores, `$`, and
-    /// comma separators for nested keys. No spaces are allowed.
-    ///
-    /// ### Returns
-    ///
-    /// * A boolean
-    fn is_valid_jsonb_search_key(&self) -> bool;
-
     /// ## Coerce the value into a boolean
     ///
     /// Accepted values are "true" and "false" (case insensitive)
@@ -88,12 +78,6 @@ impl<T: AsRef<str>> CustomStringExtensions for T {
             }
         }
         result
-    }
-
-    fn is_valid_jsonb_search_key(&self) -> bool {
-        self.as_ref()
-            .chars()
-            .all(|c| c.is_ascii_alphanumeric() || c == '_' || c == ',' || c == '$')
     }
 
     fn as_integer(&self) -> Result<Vec<i32>, ApiError> {